+++ src/System/Taffybar/DBus/Client/Util.hs
@@ -1,9 +1,11 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE TemplateHaskellQuotes #-}
 module System.Taffybar.DBus.Client.Util where
 
-import           Control.Applicative
+#if !MIN_VERSION_base(4,18,0)
+import           Control.Applicative (liftA2)
+#endif
 import           DBus.Generation
 import qualified DBus.Internal.Types as T
 import qualified DBus.Introspection as I
@@ -13,33 +15,20 @@ import           Data.Maybe
 import           Language.Haskell.TH
 import           StatusNotifier.Util (getIntrospectionObjectFromFile)
 
-#if __GLASGOW_HASKELL__ >= 802
 deriveShowAndEQ :: [DerivClause]
 deriveShowAndEQ =
   [DerivClause Nothing [ConT ''Eq, ConT ''Show]]
-#endif
 
 buildDataFromNameTypePairs :: Name -> [(Name, Type)] -> Dec
 buildDataFromNameTypePairs name pairs =
   DataD [] name [] Nothing [RecC name (map mkVarBangType pairs)]
-#if __GLASGOW_HASKELL__ >= 802
         deriveShowAndEQ
-#else
-        []
-#endif
   where mkVarBangType (fieldName, fieldType) =
           (fieldName, Bang NoSourceUnpackedness NoSourceStrictness, fieldType)
 
 
 standaloneDeriveEqShow :: Name -> [Dec]
-#if __GLASGOW_HASKELL__ < 802
-standaloneDeriveEqShow name =
-  [ StandaloneDerivD [] (ConT ''Eq `AppT` ConT name)
-  , StandaloneDerivD [] (ConT ''Show `AppT` ConT name)
-  ]
-#else
 standaloneDeriveEqShow _ = []
-#endif
 
 type GetTypeForName = String -> T.Type -> Maybe Type