+++ src/System/Taffybar/Hooks.hs
@@ -8,7 +8,8 @@
 -- Stability   : unstable
 -- Portability : unportable
 --
--- This module provides various startup hooks that can be added to 'TaffyConfig'.
+-- This module provides various startup hooks that can be added to
+-- 'TaffybarConfig'.
 -----------------------------------------------------------------------------
 
 module System.Taffybar.Hooks
@@ -58,8 +59,13 @@ setTaffyLogFormatter loggerName = do
   handler <- taffyLogHandler
   updateGlobalLogger loggerName $ setHandlers [handler]
 
--- | Add 'refreshrefreshBatteriesOnPropChange' to the 'startupHook' of the
--- provided 'TaffybarConfig'.
+-- | Add 'refreshBatteriesOnPropChange' to the 'startupHook' of the
+-- provided 'TaffybarConfig'. Use this if your system has issues with
+-- the battery widget not updating or reporting the incorrect state.
+--
+-- This function 'withBatteryRefresh' is __not normally needed__
+-- because the battery widget already subscribes to updates from
+-- UPower, and UPower usually works correctly.
 withBatteryRefresh :: TaffybarConfig -> TaffybarConfig
 withBatteryRefresh = appendHook refreshBatteriesOnPropChange