+++ src/System/Taffybar/Information/Network.hs
@@ -16,7 +16,6 @@
 
 module System.Taffybar.Information.Network where
 
-import           Control.Applicative
 import qualified Control.Concurrent.MVar as MV
 import           Control.Exception (catch, SomeException)
 import           Control.Monad
@@ -28,8 +27,6 @@ import           Data.Time.Clock.System
 import           Safe ( atMay, initSafe, readDef )
 import           System.Taffybar.Information.StreamInfo ( getParsedInfo )
 import           System.Taffybar.Util
-
-import           Prelude
 
 networkInfoFile :: FilePath
 networkInfoFile = "/proc/net/dev"
@@ -103,7 +100,7 @@ monitorNetworkInterfaces interval onUpdate = void $ do
         let speedInfo = map sampleToSpeeds samples
         onUpdate speedInfo
         return samples
-      doUpdate = MV.modifyMVar_ samplesVar ((>>= doOnUpdate) . updateSamples)
+      doUpdate = MV.modifyMVar_ samplesVar (updateSamples >=> doOnUpdate)
   foreverWithDelay interval doUpdate
 
 updateSamples ::