+++ src/System/Taffybar/Widget/Battery.hs
@@ -22,7 +22,6 @@ module System.Taffybar.Widget.Battery
   , textBatteryNewWithLabelAction
   ) where
 
-import           Control.Applicative
 import           Control.Monad
 import           Control.Monad.IO.Class
 import           Control.Monad.Trans.Reader
@@ -30,7 +29,6 @@ import           Data.Default (Default(..))
 import           Data.Int (Int64)
 import qualified Data.Text as T
 import           GI.Gtk as Gtk
-import           Prelude
 import           StatusNotifier.Tray (scalePixbufToSize)
 import           System.Taffybar.Context
 import           System.Taffybar.Information.Battery