+++ src/System/Taffybar/Widget/Crypto.hs
@@ -26,6 +26,7 @@ import           Control.Monad.Trans.Maybe
 import           Control.Monad.Trans.Reader
 import           Data.Aeson
 import           Data.Aeson.Types
+import qualified Data.Aeson.Key as Key
 import qualified Data.ByteString.Lazy as LBS
 import           Data.List.Split
 import           Data.Maybe
@@ -37,7 +38,7 @@ import qualified GI.Gtk as Gtk
 import           Network.HTTP.Simple hiding (Proxy)
 import           System.FilePath.Posix
 import           System.Taffybar.Context
-import           System.Taffybar.Information.Crypto
+import           System.Taffybar.Information.Crypto hiding (symbol)
 import           System.Taffybar.Util
 import           System.Taffybar.Widget.Generic.AutoSizeImage
 import           System.Taffybar.Widget.Generic.ChannelWidget
@@ -47,10 +48,12 @@ import           Text.Printf
 -- | Extends 'cryptoPriceLabel' with an icon corresponding to the symbol of the
 -- purchase crypto that will appear to the left of the price label. See the
 -- docstring for 'getCryptoPixbuf' for details about how this icon is retrieved.
--- As with 'cryptoPriceLabel', this function must be invoked with a type
--- application with the type string for the coinbase product that should be
--- tracked. See the docstring of 'cryptoPriceLabel' for details about the exact
--- format that this string should take.
+-- Note that automatic icon retrieval requires a coinmarketcap api key to be set
+-- at taffybar startup. As with 'cryptoPriceLabel', this function must be
+-- invoked with a type application with the type string that expresses the
+-- symbol of the relevant token and the underlying currency in which its price
+-- should be expressed. See the docstring of 'cryptoPriceLabel' for details
+-- about the exact format that this string should take.
 cryptoPriceLabelWithIcon :: forall a. KnownSymbol a => TaffyIO Gtk.Widget
 cryptoPriceLabelWithIcon = do
   label <- cryptoPriceLabel @a
@@ -67,6 +70,8 @@ cryptoPriceLabelWithIcon = do
   Gtk.containerAdd hbox image
   Gtk.containerAdd hbox label
 
+  Gtk.widgetShowAll hbox
+
   Gtk.toWidget hbox
 
 newtype CMCAPIKey = CMCAPIKey String
@@ -79,15 +84,13 @@ setCMCAPIKey :: String -> TaffyIO CMCAPIKey
 setCMCAPIKey key =
   getStateDefault $ return $ CMCAPIKey key
 
--- | Build a label that will reflect the price of some crypto product in the
--- coinbase API. A coinbase product consists of a pair of assets, the asset that
--- is being purchased, and the asset used to do the purchasing. This function
--- accepts the product to be tracked as a type parameter with kind 'String' of
--- the form `(symbol for asset being purchased)-(symbol asset used for
--- purchase)`. For example, the product string for the price of bitcoin quoted
--- in U.S. dollars is "BTC-USD". You can invoke this function by enabling the
--- TypeApplications language extension and passing the string associated with
--- the asset that you want to track as follows:
+-- | Build a label that will reflect the price of some token in some currency in
+-- the coingecko API. This function accepts these valuesas a type parameter with
+-- kind 'String' of the form `(symbol for asset being purchased)-(currency the
+-- price should be expressed in)`. For example, the product string for the price
+-- of bitcoin quoted in U.S. dollars is "BTC-USD". You can invoke this function
+-- by enabling the TypeApplications language extension and passing the string
+-- associated with the asset that you want to track as follows:
 --
 -- > cryptoPriceLabel @"BTC-USD"
 cryptoPriceLabel :: forall a. KnownSymbol a => TaffyIO Gtk.Widget
@@ -139,6 +142,4 @@ getCryptoIconFromCMC' cmcAPIKey symbol = do
 getIconURIFromJSON :: String -> LBS.ByteString -> Maybe Data.Text.Text
 getIconURIFromJSON symbol jsonText =
   decode jsonText >>= parseMaybe
-           ((.: "data") >=> (.: Data.Text.pack symbol) >=> (.: "logo"))
-
-
+           ((.: "data") >=> (.: Key.fromString symbol) >=> (.: "logo"))